+++ packages/notification/channel/instatus/index.ts
@@ -22,7 +22,7 @@
  * SOFTWARE.                                                                      *
  **********************************************************************************/
 
-import axios from 'axios'
+import axios, { AxiosError } from 'axios'
 import Joi from 'joi'
 import {
   findIncident,
@@ -91,8 +91,9 @@ export async function send(
       break
     }
 
-    default:
+    default: {
       break
+    }
   }
 }
 
@@ -124,7 +125,7 @@ async function createIncident(
     }))
     const data = {
       name: 'Service is down',
-      message: message,
+      message,
       components: [componentID],
       started,
       status,
@@ -137,10 +138,13 @@ async function createIncident(
       .then((res) => res?.data?.id)
 
     await insertIncidentToDatabase({ incidentID, probeID, status, url })
-  } catch (error: any) {
+  } catch (error: unknown) {
+    const axiosError = error instanceof AxiosError ? error : new AxiosError()
     throw new Error(
-      `${error?.message}${
-        error?.data ? `. ${error?.response?.data?.message}` : ''
+      `${axiosError?.message}${
+        axiosError?.response?.data
+          ? `. ${axiosError?.response?.data?.message}`
+          : ''
       }`
     )
   }
@@ -186,10 +190,13 @@ async function updateIncident(
       data,
       getAxiosConfig(apiKey)
     )
-  } catch (error: any) {
+  } catch (error: unknown) {
+    const axiosError = error instanceof AxiosError ? error : new AxiosError()
     throw new Error(
-      `${error?.message}${
-        error?.data ? `. ${error?.response?.data?.message}` : ''
+      `${axiosError.message}${
+        axiosError?.response?.data
+          ? `. ${axiosError?.response?.data?.message}`
+          : ''
       }`
     )
   }