+++ src/components/logger/history.ts
@@ -31,6 +31,7 @@ import { Probe } from '../../interfaces/probe'
 import type { Notification } from '@hyperjumptech/monika-notification'
 import { log } from '../../utils/pino'
 import { getConfig } from '../config'
+import { getErrorMessage } from '../../utils/catch-error-handler'
 const sqlite3 = SQLite3.verbose()
 const dbPath = path.resolve(process.cwd(), 'monika-logs.db')
 
@@ -46,15 +47,8 @@ export type UnreportedRequestsLog = {
   id: number
   timestamp: number
   probeId: string
-  probeName?: string
-  requestMethod: string
-  requestUrl: string
-  requestHeader?: string
-  requestBody?: string
   responseStatus: number
-  responseHeader?: string
   responseTime: number
-  responseSize?: number
   alerts: string[]
 }
 
@@ -98,7 +92,7 @@ type Summary = {
 
 type ProbeRequestDB = {
   id: number
-  /* eslint-disable camelcase */
+
   probe_id: string
   response_status: number
   request_url: string
@@ -109,20 +103,10 @@ type ProbeRequestDB = {
 type UnreportedProbeRequestDB = {
   alerts: string
   id: number
-  /* eslint-disable camelcase */
+
   probe_id: string
-  probe_name?: string
-  request_method: string
-  request_body?: string
-  request_header?: string
-  request_type: string
-  request_url: string
-  response_header?: string
-  response_size?: number
   response_status: number
   response_time: number
-  socket_host: string
-  socket_port: string
   timestamp: number
   /* eslint-enable camelcase */
 }
@@ -130,7 +114,7 @@ type UnreportedProbeRequestDB = {
 type UnreportedNotificationDB = {
   id: number
   timestamp: number
-  /* eslint-disable camelcase */
+
   probe_id: string
   probe_name: string
   alert_type: string
@@ -172,8 +156,8 @@ export async function openLogfile(): Promise<void> {
     setDatabase(db)
 
     await migrate()
-  } catch (error: any) {
-    log.error("Warning: Can't open logfile. " + error.message)
+  } catch (error: unknown) {
+    log.error("Warning: Can't open logfile. " + getErrorMessage(error))
   }
 }
 
@@ -193,7 +177,7 @@ export async function deleteFromProbeRequests(
   }
 
   return {
-    probeIds: probeIds,
+    probeIds,
     probeRequestIds: ids,
   }
 }
@@ -270,18 +254,8 @@ export async function getUnreportedLogs(
     SELECT PR.id,
       PR.created_at as timestamp,
       PR.probe_id,
-      PR.probe_name,
-      PR.request_method,
-      PR.request_url,
-      PR.request_header,
-      PR.request_body,
       PR.response_status,
-      PR.response_header,
       PR.response_time,
-      PR.response_size,
-      PR.request_type,
-      PR.socket_host,
-      PR.socket_port,
       CASE
         WHEN A.type IS NULL THEN json_array()
         ELSE json_group_array(A.type)
@@ -317,18 +291,8 @@ export async function getUnreportedLogs(
               alerts,
               id,
               probe_id,
-              probe_name,
-              request_body,
-              request_header,
-              request_method,
-              request_type,
-              request_url,
-              response_header,
-              response_size,
               response_status,
               response_time,
-              socket_host,
-              socket_port,
               timestamp,
             } = unreportedProbeRequest
 
@@ -336,18 +300,8 @@ export async function getUnreportedLogs(
               alerts: JSON.parse(alerts),
               id,
               probeId: probe_id,
-              probeName: probe_name || '',
-              requestBody: request_body || '',
-              requestHeader: request_header || '',
-              requestMethod: request_method,
-              requestType: request_type,
-              requestUrl: request_url,
-              responseHeader: response_header,
-              responseSize: response_size || 0,
               responseStatus: response_status,
               responseTime: response_time,
-              socketHost: socket_host,
-              socketPort: socket_port,
               timestamp,
             }
           }
@@ -533,8 +487,10 @@ export async function saveProbeRequestLog({
         ])
       )
     )
-  } catch (error: any) {
-    log.error("Error: Can't insert data into monika-log.db. " + error.message)
+  } catch (error: unknown) {
+    log.error(
+      "Error: Can't insert data into monika-log.db. " + getErrorMessage(error)
+    )
   }
 }
 
@@ -577,8 +533,10 @@ export async function saveNotificationLog(
       notification.id,
       notification.type,
     ])
-  } catch (error: any) {
-    log.error("Error: Can't insert data into monika-log.db. " + error.message)
+  } catch (error: unknown) {
+    log.error(
+      "Error: Can't insert data into monika-log.db. " + getErrorMessage(error)
+    )
   }
 }