+++ src/components/probe/prober/http/request.ts
@@ -214,7 +214,7 @@ export async function httpRequest({
           headers: '',
           responseTime,
           result: probeRequestResult.failed,
-          errMessage: (error as AxiosError).code,
+          error: (error as AxiosError).message,
         }
       }
     }
@@ -222,12 +222,12 @@ export async function httpRequest({
     // other errors
     return {
       data: '',
-      body: (error as Error).message,
+      body: '',
       status: 99,
       headers: '',
       responseTime,
       result: probeRequestResult.failed,
-      errMessage: (error as Error).message,
+      error: (error as Error).message,
     }
   }
 }
@@ -413,23 +413,36 @@ function getErrorStatusWithExplanation(error: unknown): {
       }
     }
 
-    case 'EHOSTUNREACHED': {
+    case 'EHOSTUNREACH': {
       return {
         status: 16,
-        description: 'EHOSTUNREACHED: The host is unreachable.',
+        description: 'EHOSTUNREACH: The host is unreachable.',
+      }
+    }
+
+    case 'EPROTO': {
+      return {
+        status: 17,
+        description:
+          "EPROTO: There are issues with the website's SSL/TLS certificates, incompatible protocols, or other SSL-related problems.",
       }
     }
 
     default: {
       if (error instanceof AxiosError) {
         console.error(
-          `Unhandled error while probing ${error.request.url}, got ${error.code} ${error.stack} `
+          `Error code 99: Unhandled error while probing ${error.request.url}, got ${error.code} ${error.stack} `
         )
       } else {
-        console.error(`Unhandled error, got ${(error as AxiosError).stack}`)
+        console.error(
+          `Error code 99: Unhandled error, got ${(error as AxiosError).stack}`
+        )
       }
 
-      return { status: 99, description: '' }
+      return {
+        status: 99,
+        description: `Error code 99: ${(error as AxiosError).stack}`,
+      }
     } // in the event an unlikely unknown error, send here
   }
 }