+++ src/components/probe/prober/mongo/request.ts
@@ -27,6 +27,7 @@ import * as mongodbURI from 'mongodb-uri'
 import { probeRequestResult } from '../../../../interfaces/request'
 import type { ProbeRequestResponse } from '../../../../interfaces/request'
 import { differenceInMilliseconds } from 'date-fns'
+import { getErrorMessage } from '../../../../utils/catch-error-handler'
 
 type MongoRequest = {
   uri?: string // Connection URI
@@ -71,8 +72,8 @@ export async function mongoRequest(
     baseResponse.result = probeRequestResult.success
   } else {
     baseResponse.result = probeRequestResult.failed
-    baseResponse.body = result.message
-    baseResponse.errMessage = result.message
+    baseResponse.body = ''
+    baseResponse.error = result.message
   }
 
   return baseResponse
@@ -140,8 +141,8 @@ async function sendMongoRequest(params: MongoRequest): Promise<MongoResult> {
       result.isAlive = true
       result.message = `${host}:${port} PONGED`
     }
-  } catch (error: any) {
-    result.message = error
+  } catch (error: unknown) {
+    result.message = getErrorMessage(error)
   } finally {
     if (client) {
       await client.close()