+++ src/symon/index.test.ts
@@ -25,18 +25,39 @@
 import { expect } from '@oclif/test'
 import { rest } from 'msw'
 import { setupServer } from 'msw/node'
-import sinon from 'sinon'
 
-import type { MonikaFlags } from '../flag'
+import { monikaFlagsDefaultValue, type MonikaFlags } from '../flag'
 import type { Config } from '../interfaces/config'
+import type { Probe } from '../interfaces/probe'
 
 import SymonClient from '.'
+import { getContext, resetContext, setContext } from '../context'
+import { deleteProbe, getProbes } from '../components/config/probe'
 import { validateProbes } from '../components/config/validation'
-import * as loggerHistory from '../components/logger/history'
-import { setContext } from '../context'
+import events from '../events'
+import { md5Hash } from '../utils/hash'
+import { getEventEmitter } from '../utils/events'
+import { getErrorMessage } from '../utils/catch-error-handler'
 
-let getUnreportedLogsStub: sinon.SinonStub
-
+const config: Config = {
+  version: 'asdfg123',
+  probes: [
+    {
+      id: '1',
+      name: 'test',
+      interval: 10,
+      requests: [],
+      alerts: [],
+    },
+    {
+      id: '2',
+      name: 'Example',
+      interval: 10,
+      requests: [{ url: 'https://example.com', body: '', timeout: 2000 }],
+      alerts: [],
+    },
+  ],
+}
 const server = setupServer(
   rest.get('http://ip-api.com/json/192.168.1.1', (_, res, ctx) =>
     res(
@@ -63,6 +84,17 @@ const server = setupServer(
   rest.post('http://localhost:4000/api/v1/monika/status', (_, res, ctx) =>
     res(ctx.status(200))
   ),
+  rest.get('http://localhost:4000/api/v1/monika/1234/probes', (_, res, ctx) =>
+    res(
+      ctx.set('etag', config.version || ''),
+      ctx.json({
+        statusCode: 'ok',
+        message: 'Successfully get probes configuration',
+        data: config.probes,
+      })
+    )
+  ),
+  rest.get('https://example.com', (_, res, ctx) => res(ctx.json({}))),
   rest.get('http://localhost:4000/api/v1/monika/report', (_, res, ctx) =>
     res(
       ctx.json({
@@ -78,37 +110,27 @@ describe('Symon initiate', () => {
     server.listen()
   })
   beforeEach(() => {
+    resetContext()
     setContext({
       flags: {
-        symonUrl: 'https://example.com',
+        symonUrl: 'http://localhost:4000',
         symonKey: 'random-key',
+        symonGetProbesIntervalMs:
+          monikaFlagsDefaultValue.symonGetProbesIntervalMs,
       } as MonikaFlags,
     })
 
-    getUnreportedLogsStub = sinon
-      .stub(loggerHistory, 'getUnreportedLogs')
-      .resolves({ requests: [], notifications: [] })
+    // reset probe cache
+    for (const { id } of getProbes()) {
+      deleteProbe(id)
+    }
   })
   afterEach(() => {
     server.resetHandlers()
-    getUnreportedLogsStub.restore()
   })
   after(() => {
     server.close()
   })
-
-  const config: Config = {
-    version: 'asdfg123',
-    probes: [
-      {
-        id: '1',
-        name: 'test',
-        interval: 10,
-        requests: [],
-        alerts: [],
-      },
-    ],
-  }
 
   it('should send handshake data on initiate', async () => {
     setContext({
@@ -136,18 +158,6 @@ describe('Symon initiate', () => {
             })
           )
         }
-      ),
-      rest.get(
-        'http://localhost:4000/api/v1/monika/1234/probes',
-        (_, res, ctx) =>
-          res(
-            ctx.set('etag', config.version || ''),
-            ctx.json({
-              statusCode: 'ok',
-              message: 'Successfully get probes configuration',
-              data: config.probes,
-            })
-          )
       )
     )
 
@@ -155,10 +165,8 @@ describe('Symon initiate', () => {
       symonUrl: 'http://localhost:4000',
       symonKey: 'abcd',
     })
-    const reportSpy = sinon.spy(symon, 'report')
     await symon.initiate()
-    await symon.stopReport()
-    expect(symon.monikaId).equals('1234')
+    await symon.stop()
 
     expect(body.publicIp).equals('192.168.1.1')
     expect(body.pid).greaterThan(0)
@@ -170,46 +178,28 @@ describe('Symon initiate', () => {
     expect(body.privateIp).length.greaterThan(0)
     expect(body.os).length.greaterThan(0)
     expect(body.version).equals('v1.5.0')
-
-    expect(reportSpy.called).equals(true)
   }).timeout(15_000)
 
   it('should fetch probes config on initiate', async () => {
-    server.use(
-      rest.get(
-        'http://localhost:4000/api/v1/monika/1234/probes',
-        (_, res, ctx) =>
-          res(
-            ctx.set('etag', config.version || ''),
-            ctx.json({
-              statusCode: 'ok',
-              message: 'Successfully get probes configuration',
-              data: config.probes,
-            })
-          )
-      )
-    )
-
     const symon = new SymonClient({
       symonUrl: 'http://localhost:4000',
       symonKey: 'abcd',
     })
-    sinon.spy(symon, 'report')
 
-    expect(symon.config).to.be.null
+    expect(getContext().config).to.be.undefined
 
     await symon.initiate()
-    await symon.stopReport()
+    await symon.stop()
 
-    expect(symon.config).deep.equals({
+    expect(getContext().config).deep.equals({
       ...config,
       probes: await validateProbes(config.probes),
     })
+    expect(getProbes()).deep.eq(await validateProbes(config.probes))
   }).timeout(15_000)
 
   it('should throw an error if the request to get probes is failed', async () => {
     // arrange
-    server.resetHandlers()
     server.use(
       rest.post(
         'http://localhost:4000/api/v1/monika/client-handshake',
@@ -223,7 +213,10 @@ describe('Symon initiate', () => {
               },
             })
           )
-      )
+      ),
+      rest.get('http://localhost:4000/api/v1/monika/1234/probes', () => {
+        throw new Error('Failed')
+      })
     )
 
     const symon = new SymonClient({
@@ -235,9 +228,11 @@ describe('Symon initiate', () => {
     try {
       // act
       await symon.initiate()
-    } catch (error: any) {
-      errorMessage = error?.message
+    } catch (error: unknown) {
+      errorMessage = getErrorMessage(error)
     }
+
+    await symon.stop()
 
     // assert
     expect(errorMessage).eq('Failed to get probes from Symon')
@@ -261,25 +256,229 @@ describe('Symon initiate', () => {
         }
       )
     )
-
     const symon = new SymonClient({
       symonUrl: 'http://localhost:4000',
       symonKey: 'abcd',
+      symonMonikaId: '1234',
     })
-    sinon.spy(symon, 'report')
-    symon.monikaId = '1234'
+    await symon.initiate()
 
     // act
-    await symon.notifyEvent({
-      event: 'incident',
-      alertId: 'alert86',
-      response: { status: 400, time: 1000 },
+    getEventEmitter().emit(events.probe.notification.willSend, {
+      probeState: 'DOWN',
+      validation: {
+        response: {
+          headers: '',
+          status: 400,
+          responseTime: 1000,
+        },
+      },
+      alertId: 'p422i',
     })
+    await symon.stop()
 
     // assert
     expect(body.monikaId).equals('1234')
     expect(body.event).equals('incident')
-    expect(body.alertId).equals('alert86')
-    expect(body.response).deep.equals({ status: 400, time: 1000 })
+    expect(body.alertId).equals('p422i')
+    expect(body.response).deep.equals({
+      headers: {},
+      status: 400,
+      time: 1000,
+    })
+  }).timeout(15_000)
+
+  it('should add a new probe', async () => {
+    // arrange
+    const symonGetProbesIntervalMs = 100
+    setContext({
+      flags: {
+        symonUrl: 'http://localhost:4000',
+        symonKey: 'random-key',
+        symonGetProbesIntervalMs,
+      } as MonikaFlags,
+    })
+    const symon = new SymonClient({
+      symonUrl: 'http://localhost:4000',
+      symonKey: 'abcd',
+    })
+
+    // 1. Check initial probe cache
+    expect(getProbes()).deep.eq([])
+
+    // act
+    // 2. Initiate Symon and get all the probes
+    await symon.initiate()
+
+    // assert
+    // 3. Check the probe data after connected to Symon
+    expect(getProbes()).deep.eq(await validateProbes(config.probes))
+
+    // arrange
+    // 4. Simulate adding a probe
+    const newProbe: Probe = {
+      id: '3',
+      name: 'New Probe',
+      interval: 2,
+      requests: [{ url: 'https://example.com', body: '', timeout: 2000 }],
+      alerts: [],
+    }
+    server.use(
+      rest.get(
+        'http://localhost:4000/api/v1/monika/1234/probes',
+        (_, res, ctx) => {
+          const newProbes: Probe[] = [...config.probes, newProbe]
+
+          return res(
+            ctx.set('etag', md5Hash(newProbes)),
+            ctx.json({
+              statusCode: 'ok',
+              message: 'Successfully get probes configuration',
+              data: newProbes,
+            })
+          )
+        }
+      )
+    )
+
+    // act
+    // 5. Wait for the probe fetch to run
+    await sleep(symonGetProbesIntervalMs)
+
+    // assert
+    // 6. Check the updated probe cache
+    expect(getProbes()).deep.eq(
+      await validateProbes([...config.probes, newProbe])
+    )
+
+    // act
+    // 7. Wait for probe fetch to run
+    await sleep(symonGetProbesIntervalMs)
+
+    // assert
+    // 8. Should not update the probe cache
+    expect(getProbes()).deep.eq(
+      await validateProbes([...config.probes, newProbe])
+    )
+    await symon.stop()
+  }).timeout(15_000)
+
+  it('should update a probe', async () => {
+    // arrange
+    const symonGetProbesIntervalMs = 100
+    setContext({
+      flags: {
+        symonUrl: 'http://localhost:4000',
+        symonKey: 'random-key',
+        symonGetProbesIntervalMs,
+      } as MonikaFlags,
+    })
+    const symon = new SymonClient({
+      symonUrl: 'http://localhost:4000',
+      symonKey: 'abcd',
+    })
+
+    // 1. Check initial probe cache
+    expect(getProbes()).deep.eq([])
+
+    // act
+    // 2. Initiate Symon and get all the probes
+    await symon.initiate()
+
+    // assert
+    // 3. Check the probe data after connected to Symon
+    expect(getProbes()).deep.eq(await validateProbes(config.probes))
+
+    // arrange
+    // 4. Simulate updating a probe
+    const updatedProbes: Probe[] = [{ ...config.probes[0], interval: 5 }]
+    server.use(
+      rest.get(
+        'http://localhost:4000/api/v1/monika/1234/probes',
+        (_, res, ctx) => {
+          const newProbes: Probe[] = updatedProbes
+
+          return res(
+            ctx.set('etag', md5Hash(newProbes)),
+            ctx.json({
+              statusCode: 'ok',
+              message: 'Successfully get probes configuration',
+              data: newProbes,
+            })
+          )
+        }
+      )
+    )
+
+    // act
+    // 5. Wait for the probe fetch to run
+    await sleep(symonGetProbesIntervalMs)
+
+    // assert
+    // 6. Check the updated probe cache
+    expect(getProbes()).deep.eq(await validateProbes(updatedProbes))
+
+    await symon.stop()
+  }).timeout(15_000)
+
+  it('should delete a probe', async () => {
+    // arrange
+    const symonGetProbesIntervalMs = 100
+    setContext({
+      flags: {
+        symonUrl: 'http://localhost:4000',
+        symonKey: 'random-key',
+        symonGetProbesIntervalMs,
+      } as MonikaFlags,
+    })
+    const symon = new SymonClient({
+      symonUrl: 'http://localhost:4000',
+      symonKey: 'abcd',
+    })
+
+    // 1. Check initial probe cache
+    expect(getProbes()).deep.eq([])
+
+    // act
+    // 2. Initiate Symon and get all the probes
+    await symon.initiate()
+
+    // assert
+    // 3. Check the probe data after connected to Symon
+    expect(getProbes()).deep.eq(await validateProbes(config.probes))
+
+    // arrange
+    // 4. Simulate deleting a probe
+    const updatedProbes: Probe[] = config.probes.filter(({ id }) => id === '1')
+    server.use(
+      rest.get(
+        'http://localhost:4000/api/v1/monika/1234/probes',
+        (_, res, ctx) =>
+          res(
+            ctx.set('etag', md5Hash(updatedProbes)),
+            ctx.json({
+              statusCode: 'ok',
+              message: 'Successfully get probes configuration',
+              data: updatedProbes,
+            })
+          )
+      )
+    )
+
+    // act
+    // 5. Wait for the probe fetch to run
+    await sleep(symonGetProbesIntervalMs)
+
+    // assert
+    // 6. Check the updated probe cache
+    expect(getProbes()).deep.eq(await validateProbes(updatedProbes))
+
+    await symon.stop()
   }).timeout(15_000)
 })
+
+function sleep(durationMs: number): Promise<void> {
+  return new Promise((resolve) => {
+    setTimeout(resolve, durationMs)
+  })
+}